Write results to file not to stdandard output
With this change all the results will be aggregated to a file instead of
printing them to stdout. This simplifies performance indicators handling that
were printing to stound/stderr from a forked process causing serialization
issues when the output was expected to be a JSON output (printed logs bloated
output). Also, with this fix we do not rely on OpenShift to store logs - these
logs were automatically deleted after some time causing troubles aggregating
them when they were no longer available.